refactor: restructure server files to divide by purpose & better support multiple rendering targets
