+++ dotcom-rendering/src/server/dev-server.ts
@@ -1,16 +1,15 @@
 import type { Handler } from 'express';
-import { handleAMPArticle } from '../amp/server';
-import { handleAppsArticle } from '../apps/server';
+import { handleAMPArticle } from '../amp/server/index.article';
+import { handleAppsArticle } from '../apps/server/index.article';
+import { handleAllEditorialNewslettersPage } from '../web/server/index.allEditorialNewslettersPage';
 import {
-	handleAllEditorialNewslettersPage,
 	handleArticle,
 	handleArticleJson,
 	handleBlocks,
-	handleFront,
-	handleFrontJson,
 	handleInteractive,
 	handleKeyEvents,
-} from '../web/server';
+} from '../web/server/index.article';
+import { handleFront, handleFrontJson } from '../web/server/index.front';
 
 /** article URLs contain a part that looks like “2022/nov/25” */
 const ARTICLE_URL = /\/\d{4}\/[a-z]{3}\/\d{2}\//;
